Print changes and support for flags and versions
All user-intended print messages that aren't the "done" print at the end now begin with "ERROR" or "WARNING" to let the user know they messed up.

Added support to read versions and a basic check to see if inputted file is actually a FEF.

Also added flag support. Inputting a file must be preceeded by "-i". "-f" flag yet to be added.

Finally, a miniscule change to the readme.