Multiple unnamed file scenario averted
Added a default name and fixed a crash for when files do not have a name. This should never happen in the first place, but you never know.

Default file name was already partially implemented, but this now handles if there are multiple unnamed files so they don't overwrite each other.