Add output folder
Files will now go to the output folder, which is by default the name of the inputted FEF file. Users can change this using the -f flag, where they can put a directory. A directory that does not clearly begin with a relative marker or have a drive (on Windows) marker will be put in the current directory.